requirements.txt
dj-database-url==0.5.0
Django==2.1.4
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-filter==2.0.0
django-heroku==0.3.1
django-markdown-deux==1.0.5
django-pagedown==1.0.6
djangorestframework==3.9.0
djangorestframework-bulk==0.2.1
gunicorn==19.9.0
Markdown==3.0.1
markdown2==2.3.7
Pillow==5.3.0
psycopg2==2.7.6.1
pytz==2018.7
whitenoise==4.1.2
pylint
black
